requirements.txt
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.3.1
prompt-toolkit==3.0.38
PyJWT==2.7.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
psycopg2-binary==2.9.6
flower==2.0.0
humanize==4.7.0
prometheus-client==0.17.1
tornado==6.3.2
